+++ web-common/src/features/entity-management/WatchFilesClient.ts
@@ -6,10 +6,14 @@ import {
   V1FileEvent,
   V1WatchFilesResponse,
 } from "@rilldata/web-common/runtime-client";
-import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
+import {
+  runtime,
+  projectInitialized,
+} from "@rilldata/web-common/runtime-client/runtime-store";
 import { WatchRequestClient } from "@rilldata/web-common/runtime-client/watch-request-client";
 import { get } from "svelte/store";
 import { queryClient } from "@rilldata/web-common/lib/svelte-query/globalQueryClient";
+import { goto } from "$app/navigation";
 
 export class WatchFilesClient {
   public readonly client: WatchRequestClient<V1WatchFilesResponse>;
@@ -30,7 +34,7 @@ export class WatchFilesClient {
     });
   }
 
-  private handleWatchFileResponse(res: V1WatchFilesResponse) {
+  private async handleWatchFileResponse(res: V1WatchFilesResponse) {
     if (!res?.path || res.path.includes(".db")) return;
 
     const instanceId = get(runtime).instanceId;
@@ -45,11 +49,17 @@ export class WatchFilesClient {
             getRuntimeServiceGetFileQueryKey(instanceId, { path: res.path }),
           );
           void fileArtifacts.fileUpdated(res.path);
+
           if (res.path === "/rill.yaml") {
             // If it's a rill.yaml file, invalidate the dev JWT queries
             void queryClient.invalidateQueries(
               getRuntimeServiceIssueDevJWTQueryKey(),
             );
+
+            if (!get(projectInitialized)) {
+              projectInitialized.set(true);
+              if (window.location.pathname === "/welcome") await goto("/");
+            }
           }
           this.seenFiles.add(res.path);
           break;
@@ -60,6 +70,11 @@ export class WatchFilesClient {
           );
           fileArtifacts.fileDeleted(res.path);
           this.seenFiles.delete(res.path);
+
+          if (res.path === "/rill.yaml") {
+            projectInitialized.set(false);
+            await goto("/welcome");
+          }
           break;
       }
     }
