+++ web-common/src/features/sources/modal/FileDrop.svelte
@@ -8,17 +8,16 @@
   import { useQueryClient } from "@tanstack/svelte-query";
   import { runtime } from "../../../runtime-client/runtime-store";
   import { EMPTY_PROJECT_TITLE } from "../../welcome/constants";
-  import { useIsProjectInitialized } from "../../welcome/is-project-initialized";
   import { compileCreateSourceYAML } from "../sourceUtils";
   import { createSource } from "./createSource";
   import { uploadTableFiles } from "./file-upload";
+  import { projectInitialized } from "../../../runtime-client/runtime-store";
 
   export let showDropOverlay: boolean;
 
   const queryClient = useQueryClient();
 
   $: runtimeInstanceId = $runtime.instanceId;
-  $: isProjectInitialized = useIsProjectInitialized(runtimeInstanceId);
 
   const unpackEmptyProject = createRuntimeServiceUnpackEmpty();
 
@@ -37,7 +36,7 @@
     for await (const { tableName, filePath } of uploadedFiles) {
       try {
         // If project is uninitialized, initialize an empty project
-        if (!$isProjectInitialized.data) {
+        if (!$projectInitialized) {
           $unpackEmptyProject.mutate({
             instanceId: $runtime.instanceId,
             data: {
