+++ web-common/src/features/sources/modal/LocalSourceUpload.svelte
@@ -14,16 +14,14 @@
   import { createEventDispatcher } from "svelte";
   import { runtime } from "../../../runtime-client/runtime-store";
   import { EMPTY_PROJECT_TITLE } from "../../welcome/constants";
-  import { useIsProjectInitialized } from "../../welcome/is-project-initialized";
   import { compileCreateSourceYAML } from "../sourceUtils";
   import { createSource } from "./createSource";
+  import { projectInitialized } from "../../../runtime-client/runtime-store";
 
   const dispatch = createEventDispatcher();
   const queryClient = useQueryClient();
 
   $: runtimeInstanceId = $runtime.instanceId;
-
-  $: isProjectInitialized = useIsProjectInitialized(runtimeInstanceId);
 
   const unpackEmptyProject = createRuntimeServiceUnpackEmpty();
 
@@ -36,7 +34,7 @@
     for await (const { tableName, filePath } of uploadedFiles) {
       try {
         // If project is uninitialized, initialize an empty project
-        if (!$isProjectInitialized.data) {
+        if (!$projectInitialized) {
           $unpackEmptyProject.mutate({
             instanceId: $runtime.instanceId,
             data: {
