+++ web-local/src/routes/(viz)/custom/[name]/+page.ts
@@ -4,9 +4,11 @@ import type { QueryFunction } from "@tanstack/svelte-query";
 import { error } from "@sveltejs/kit";
 import { ResourceKind } from "@rilldata/web-common/features/entity-management/resource-selectors.js";
 import { queryClient } from "@rilldata/web-common/lib/svelte-query/globalQueryClient.js";
+import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
+import { get } from "svelte/store";
 
-export const load = async ({ parent, params, depends }) => {
-  const { instanceId } = await parent();
+export const load = async ({ params, depends }) => {
+  const { instanceId } = get(runtime);
 
   const dashboardName = params.name;
